refactor: limit casting to full variant only to avoid GMS dependencies for F-Droid
