+++ app/src/fdroid/kotlin/me/echeung/moemoekyun/ui/common/CastButton.kt
@@ -1,0 +1,12 @@
+package me.echeung.moemoekyun.ui.common
+
+import androidx.compose.runtime.Composable
+
+/**
+ * Cast button stub for F-Droid variant.
+ * Casting functionality is not available in F-Droid builds.
+ */
+@Composable
+fun CastButton() {
+    // No-op: Cast functionality is only available in the full variant
+}
