+++ app/src/main/kotlin/me/echeung/moemoekyun/di/MediaModule.kt
@@ -3,12 +3,8 @@ package me.echeung.moemoekyun.di
 import android.content.Context
 import android.media.AudioManager
 import androidx.annotation.OptIn
-import androidx.media3.cast.CastPlayer
-import androidx.media3.cast.DefaultMediaItemConverter
-import androidx.media3.cast.RemoteCastPlayer
 import androidx.media3.common.AudioAttributes
 import androidx.media3.common.C
-import androidx.media3.common.Player
 import androidx.media3.common.util.UnstableApi
 import androidx.media3.datasource.DefaultDataSource
 import androidx.media3.datasource.DefaultHttpDataSource
@@ -21,10 +17,12 @@ import dagger.Reusable
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ServiceComponent
 import dagger.hilt.android.qualifiers.ApplicationContext
-import me.echeung.moemoekyun.service.PlaybackMediaItemConverter
 import me.echeung.moemoekyun.util.ext.audioManager
 import me.echeung.moemoekyun.util.system.NetworkUtil
 
+/**
+ * Common media module providing dependencies shared by all variants.
+ */
 @Module
 @OptIn(UnstableApi::class)
 @InstallIn(ServiceComponent::class)
@@ -40,15 +38,13 @@ object MediaModule {
         .setUsage(C.USAGE_MEDIA)
         .build()
 
-    @OptIn(UnstableApi::class)
     @Provides
-    fun dateSourceFactory(@ApplicationContext context: Context): DefaultDataSource.Factory = DefaultDataSource.Factory(
+    fun dataSourceFactory(@ApplicationContext context: Context): DefaultDataSource.Factory = DefaultDataSource.Factory(
         context,
         DefaultHttpDataSource.Factory()
             .setUserAgent(NetworkUtil.userAgent),
     )
 
-    @OptIn(UnstableApi::class)
     @Provides
     fun progressiveMediaSourceFactory(dataSourceFactory: DefaultDataSource.Factory): ProgressiveMediaSource.Factory =
         ProgressiveMediaSource.Factory(dataSourceFactory, DefaultExtractorsFactory())
@@ -62,26 +58,5 @@ object MediaModule {
         .setMediaSourceFactory(progressiveMediaSourceFactory)
         .setAudioAttributes(audioAttributes, true)
         .setWakeMode(C.WAKE_MODE_NETWORK)
-        .build()
-
-    @Provides
-    fun defaultMediaItemConverter(): DefaultMediaItemConverter = DefaultMediaItemConverter()
-
-    @Provides
-    fun remotePlayer(
-        @ApplicationContext context: Context,
-        mediaItemConverter: PlaybackMediaItemConverter,
-    ): RemoteCastPlayer = RemoteCastPlayer.Builder(context)
-        .setMediaItemConverter(mediaItemConverter)
-        .build()
-
-    @Provides
-    fun castPlayer(
-        @ApplicationContext context: Context,
-        exoPlayer: ExoPlayer,
-        remotePlayer: RemoteCastPlayer,
-    ): Player = CastPlayer.Builder(context)
-        .setLocalPlayer(exoPlayer)
-        .setRemotePlayer(remotePlayer)
         .build()
 }
