+++ gradle/libs.versions.toml
@@ -91,7 +91,7 @@ coil = ["coil-core", "coil-compose", "coil-gif", "coil-network-okhttp"]
 compose = ["compose-activity", "compose-material", "compose-material-icons-extended", "compose-material3", "compose-ui", "compose-ui-tooling", "compose-ui-util"]
 coroutines = ["coroutines-core", "coroutines-android"]
 junit-runtime = ["junit-platform-launcher", "junit-jupiter-engine"]
-media = ["androidx-media3-cast", "androidx-media3-common", "androidx-media3-exoplayer", "androidx-media3-session"]
+media = ["androidx-media3-common", "androidx-media3-exoplayer", "androidx-media3-session"]
 okhttp = ["okhttp-core", "okhttp-logging"]
 navigation = ["navigation3-runtime", "navigation3-ui", "androidx-lifecycle-viewmodel-navigation3", "hilt-navigation-compose"]
 preferences = ["preferences-androidx", "preferences-flow"]
